services/query_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
weaviate-client>=3.26,<4.0
requests
pyvi
PyJWT
